Update packaging, tests to test packaging, and package version
