+++ .github/workflows/linux-test.yml
@@ -73,4 +73,5 @@ jobs:
 
     - name: Run tests with pytest
       run: |
+        pip install .
         pytest
