--- setup.py
@@ -1,12 +1,0 @@
-# -*- coding: utf-8 -*-
-"""
-Created on Wed Jun 30 11:22:08 2021
-
-@author: pearsonra
-"""
-import setuptools
-
-setuptools.setup(
-    package_dir={"": "src"},
-    packages=setuptools.find_packages(where="src"),
-)
