+++ src/geoapis/__init__.py
@@ -4,4 +4,4 @@ Created on Thu Jul  1 09:57:30 2021
 
 @author: pearsonra
 """
-__version__ = "0.3.3"
+__version__ = "0.3.4"
