+++ tests/test_vector_linz_in_bounds/test.py
@@ -14,7 +14,7 @@ import shutil
 import dotenv
 import os
 
-from src.geoapis import vector
+from geoapis import vector
 
 
 class LinzVectorsTest(unittest.TestCase):
