+++ tests/test_vector_lris/test.py
@@ -13,7 +13,7 @@ import dotenv
 import os
 import geopandas
 
-from src.geoapis import vector
+from geoapis import vector
 
 
 class LrisVectorsTest(unittest.TestCase):
