+++ tests/test_vector_stats_nz/test.py
@@ -8,32 +8,29 @@ Created on Wed Jun 30 11:11:25 2021
 import unittest
 import json
 import pathlib
-import shapely
-import geopandas
 import shutil
 import dotenv
 import os
+import geopandas
 
-from src.geoapis import vector
+from geoapis import vector
 
 
 class StatsNzVectorsTest(unittest.TestCase):
-    """A class to test the basic vector.StatsNz functionality by downloading files from the dataservice within a
-    small region. The vector attributes are then compared against the expected.
+    """A class to test the basic vector.StatsNz functionality by downloading files from the dataservice.
+    The vector attributes are then compared against the expected.
 
     Tests run include (test_#### indicates the layer tested):
-        * test_105133 - Test the specified layer features are correctly downloaded within the specified bbox
-        * test_105133_no_geometry_name - Test the specified layer and bbox, but with no geometry_name given
-        * test_87883 - Test the specified layer features are correctly downloaded within the specified bbox
-        * test_87883_no_geometry_name - Test the specified layer and bbox, but with no geometry_name given
+        * test_105133 - Test the specified layer features are correctly downloaded
+        * test_8347 - Test the specified layer features are correctly downloaded
     See the associated description for keywords that can be used to search for the layer in the data service.
     """
 
     # The expected datasets and files to be downloaded - used for comparison in the later tests
     REGIONAL_COUNCILS = {
-        "area": 15945315588.69047,
+        "area": 428759923712.16785,
         "geometryType": "Polygon",
-        "length": 579211.9585905309,
+        "length": 15946566.116020141,
         "columns": [
             "geometry",
             "REGC2021_V1_00",
@@ -44,12 +41,18 @@ class StatsNzVectorsTest(unittest.TestCase):
             "Shape_Length",
             "Shape_Area",
         ],
-        "AREA_SQ_KM": [15945.3207507],
+        "AREA_SQ_KM": [
+            30084.2732362,
+            16156.2062737,
+            34888.8317055,
+            21883.7483717,
+            13989.049873,
+        ],
     }
     DISTRICT_HEALTH_BOARDS = {
-        "area": 136597398950.17044,
+        "area": 428759924063.5825,
         "geometryType": "MultiPolygon",
-        "length": 15794952.293961357,
+        "length": 34998519.751260854,
         "columns": [
             "geometry",
             "DHB2015_Code",
@@ -57,7 +60,7 @@ class StatsNzVectorsTest(unittest.TestCase):
             "Shape_Length",
             "Shape_Area",
         ],
-        "DHB2015_Code": ["14", "99"],
+        "DHB2015_Code": ["01", "02", "03", "04", "05"],
     }
 
     @classmethod
@@ -67,7 +70,7 @@ class StatsNzVectorsTest(unittest.TestCase):
 
         # load in the test instructions
         file_path = pathlib.Path().cwd() / pathlib.Path(
-            "tests/test_vector_stats_nz_in_bounds/instruction.json"
+            "tests/test_vector_stats_nz/instruction.json"
         )
         with open(file_path, "r") as file_pointer:
             cls.instructions = json.load(file_pointer)
@@ -87,34 +90,21 @@ class StatsNzVectorsTest(unittest.TestCase):
             shutil.rmtree(cls.cache_dir)
         cls.cache_dir.mkdir()
 
-        # create fake catchment boundary
-        x0 = 1752000
-        x1 = 1753000
-        y0 = 5430000
-        y1 = 5440000
-        catchment = shapely.geometry.Polygon([(x0, y0), (x0, y1), (x1, y1), (x1, y0)])
-        catchment = geopandas.GeoSeries([catchment])
-        catchment = catchment.set_crs(cls.instructions["instructions"]["projection"])
-
-        # save faked catchment file
-        catchment_dir = cls.cache_dir / "catchment"
-        catchment.to_file(catchment_dir)
-        shutil.make_archive(
-            base_name=catchment_dir, format="zip", root_dir=catchment_dir
-        )
-        shutil.rmtree(catchment_dir)
-
-        # cconvert catchment file to zipfile
-        catchment_dir = pathlib.Path(str(catchment_dir) + ".zip")
-        catchment_polygon = geopandas.read_file(catchment_dir)
-        cls.catchment_polygon = catchment_polygon
-
         # Run pipeline - download files
         cls.runner = vector.StatsNz(
             cls.instructions["instructions"]["apis"]["stats_nz"]["key"],
-            crs=None,
-            bounding_polygon=catchment_polygon,
+            crs=cls.instructions["instructions"]["projection"],
+            bounding_polygon=None,
             verbose=True,
+        )
+
+        cls.runner_generic = vector.WfsQuery(
+            key=cls.instructions["instructions"]["apis"]["stats_nz"]["key"],
+            crs=cls.instructions["instructions"]["projection"],
+            bounding_polygon=None,
+            verbose=True,
+            netloc_url="datafinder.stats.govt.nz",
+            geometry_names=["GEOMETRY", "shape"],
         )
 
     @classmethod
@@ -170,24 +160,7 @@ class StatsNzVectorsTest(unittest.TestCase):
         )
 
     def test_105133(self):
-        """Test expected features of layer loaded"""
-
-        features = self.runner.run(
-            self.instructions["instructions"]["apis"]["stats_nz"]["regional_councils"][
-                "layers"
-            ][0],
-            self.instructions["instructions"]["apis"]["stats_nz"]["regional_councils"][
-                "geometry_name"
-            ],
-        )
-        description = "Regional Council 2021"
-        benchmark = self.REGIONAL_COUNCILS
-
-        # check various shape attributes match those expected
-        self.compare_to_benchmark(features, benchmark, description, "AREA_SQ_KM")
-
-    def test_105133_no_geometry_name(self):
-        """Test expected features of layer loaded without specifying the geometry_name"""
+        """Test expected entire layer loaded correctly"""
 
         features = self.runner.run(
             self.instructions["instructions"]["apis"]["stats_nz"]["regional_councils"][
@@ -200,16 +173,13 @@ class StatsNzVectorsTest(unittest.TestCase):
         # check various shape attributes match those expected
         self.compare_to_benchmark(features, benchmark, description, "AREA_SQ_KM")
 
-    def test_87883(self):
-        """Test expected features of layer loaded"""
+    def test_8347(self):
+        """Test expected entire layer loaded correctly"""
 
         features = self.runner.run(
             self.instructions["instructions"]["apis"]["stats_nz"][
                 "district_health_board"
-            ]["layers"][0],
-            self.instructions["instructions"]["apis"]["stats_nz"][
-                "district_health_board"
-            ]["geometry_name"],
+            ]["layers"][0]
         )
         description = "District Health Board 2015"
         benchmark = self.DISTRICT_HEALTH_BOARDS
@@ -217,10 +187,24 @@ class StatsNzVectorsTest(unittest.TestCase):
         # check various shape attributes match those expected
         self.compare_to_benchmark(features, benchmark, description, "DHB2015_Code")
 
-    def test_87883_no_geometry_name(self):
-        """Test expected features of layer loaded without specifying the geometry_name"""
+    def test_105133_generic(self):
+        """Test expected entire layer loaded correctly"""
 
-        features = self.runner.run(
+        features = self.runner_generic.run(
+            self.instructions["instructions"]["apis"]["stats_nz"]["regional_councils"][
+                "layers"
+            ][0]
+        )
+        description = "Regional Council 2021"
+        benchmark = self.REGIONAL_COUNCILS
+
+        # check various shape attributes match those expected
+        self.compare_to_benchmark(features, benchmark, description, "AREA_SQ_KM")
+
+    def test_8347_generic(self):
+        """Test expected entire layer loaded correctly"""
+
+        features = self.runner_generic.run(
             self.instructions["instructions"]["apis"]["stats_nz"][
                 "district_health_board"
             ]["layers"][0]
